secp256k1: Move naf func to curve file.
This moves the function designed to calculate the non-adjacent form of a
scalar value to the new file designed to house curve operations and
finally deletes the btcec.go file since it was the last remaining item
in it.

It also renames the associated btcec_test.go to curve_test.go to reflect
reality.

This is continued work towards making the internals of the package
entirely independent of the crypto/elliptic and crypto/ecdsa stdlib
interfaces.
