+++ dcrec/secp256k1/curve_test.go
@@ -1,6 +1,4 @@
-// Copyright 2011 The Go Authors. All rights reserved.
 // Copyright (c) 2015-2020 The Decred developers
-// Copyright 2011 ThePiachu. All rights reserved.
 // Copyright 2013-2016 The btcsuite developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
